+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.3
+
+* __Adds__ Express.js compatible error handling middleware to log error stacks passed on by route request handlers.
+* __Updates__ License year.
+
 # v0.2
 
 * __Adds__ `fullPath` option. When `true`, messages logged are supplemented by caller's full path rather than just "base name" (module's file name). Defaults to `false`.